+++ tests/test_transfer_learning.py
@@ -1,0 +1,63 @@
+"""Tests for transfer-learning."""
+
+from copy import deepcopy
+
+import pandas as pd
+import pytest
+
+from baybe import Campaign
+from baybe.objectives import SingleTargetObjective
+from baybe.parameters import NumericalContinuousParameter, TaskParameter
+from baybe.recommenders import (
+    BotorchRecommender,
+    RandomRecommender,
+    TwoPhaseMetaRecommender,
+)
+from baybe.searchspace import SearchSpace
+from baybe.surrogates import GaussianProcessSurrogate
+from baybe.targets import NumericalTarget
+from baybe.utils.interval import Interval
+
+
+# @pytest.mark.parametrize("task_stratified_outtransform", [True, False])
+# @pytest.mark.parametrize("observed_test_data", [True, False])
+@pytest.mark.parametrize(
+    "task_stratified_outtransform,observed_test_data",
+    ([True, True], [False, True], [False, True]),
+)
+def test_recommendation(task_stratified_outtransform: bool, observed_test_data: bool):
+    """Test a BO iteration with multi-task model."""
+    objective = SingleTargetObjective(target=NumericalTarget(name="y", mode="MAX"))
+    parameters = [
+        NumericalContinuousParameter(name="x", bounds=Interval(0, 10)),
+        TaskParameter(name="task", values=("A", "B"), active_values=("A",)),
+    ]
+    searchspace = SearchSpace.from_product(parameters=parameters)
+    lookup = pd.DataFrame(
+        {
+            "x": [1.0, 2.0, 3.0, 4.0],
+            "y": [1.0, 2.0, 3.0, 4.0],
+            "task": ["A", "A", "B", "B"] if observed_test_data else ["B"] * 4,
+        }
+    )
+    campaign = deepcopy(
+        Campaign(
+            searchspace=searchspace,
+            objective=objective,
+            recommender=TwoPhaseMetaRecommender(
+                recommender=BotorchRecommender(
+                    surrogate_model=GaussianProcessSurrogate(
+                        task_stratified_outtransform=task_stratified_outtransform
+                    )
+                ),
+                initial_recommender=RandomRecommender(),
+            ),
+        )
+    )
+    campaign.add_measurements(lookup)
+    _ = campaign.recommend(batch_size=1)
+
+
+# TODO once stratified standardise works with recommendation
+#  without data for active task is fixed
+#  add the missing test where no active task data is added at the start